refactor

+++ findingWater/View/MapTest.swift
@@ -32,22 +32,38 @@ struct MapTest: View {
     @EnvironmentObject var viewModel2 : CreateAccount
     @State   var value:Bool = false
     var body: some View {
-        
-        NavigationView{
-          
-
+       
+            NavigationView{
+                
                 ScrollView {
                     LazyVGrid(columns: columns, spacing: 20) {
+                        Button{
+                            if viewModel2.userSession != nil  {
+                                 viewModel.createRestaurant()
+                                 viewModel2.currentUser?.value.toggle()
+                                 viewModel2.EditProducti(id: viewModel2.currentUser?.id  ?? "reewwedsewd", value: viewModel2.currentUser?.value ?? true)
+                             }
+                        } label: {
+                            Image(systemName: "plus")
+                                .foregroundColor(.white)
+                                .padding()
+                                .background(.green)
+                                .clipShape(Circle())
+                        }
+                        .disabled( viewModel2.currentUser?.value == true)
                         ForEach(viewModel.location, id: \.self) { item in
-                            if viewModel2.userSession != nil {
+                           
                                 Button{
-                                    
-                                    guard let uid = Auth.auth().currentUser?.uid else { return }
-                                    
-                                    self.value.toggle()
-                                    //viewModel.EditProducti(id: user.id ?? "JK3REWDCM", value: value)
-                                    viewModel.excluirDaAgenda(documento: item.id ?? "brecdbjbewdbn" )
-                                    
+                                    if viewModel2.userSession != nil &&   viewModel2.currentUser?.value == true {
+                                        guard (Auth.auth().currentUser?.uid) != nil else { return }
+                                        
+                                        viewModel2.currentUser?.value.toggle()
+                                        
+                                        viewModel.excluirDaAgenda(documento: item.id ?? "brecdbjbewdbn" )
+                                        
+                                        viewModel2.EditProducti(id: viewModel2.currentUser?.id  ?? "reewwedsewd", value: viewModel2.currentUser?.value ?? false)
+                                        
+                                    }
                                     
                                 } label: {
                                     Image(systemName: "bolt.fill")
@@ -57,21 +73,7 @@ struct MapTest: View {
                                         .clipShape(Circle())
                                     
                                 }
-                            } else {
-                                
-                                Button{
-                                    
-                                    
-                                    
-                                }label: {
-                                    Image(systemName: "bolt.fill")
-                                        .foregroundColor(.white)
-                                        .padding()
-                                        .background(.red)
-                                        .clipShape(Circle())
-                                }
-                            }
-                            
+                             
                             
                             
                             
@@ -82,37 +84,47 @@ struct MapTest: View {
                     }
                 }
                 
+                
+                
+                
+            }
+            .onAppear{
+                self.viewModel.fetchData()
+                self.viewModel2.fetchUser()
+                viewModel2.userSession
+                
+                var name = viewModel2.currentUser?.name
+                print(name)
                
+                print(viewModel2.currentUser?.value )
+            }
             
-            
-        }
-        .onAppear{
-            self.viewModel.fetchData()
-            self.viewModel2.fetchUser()
-            viewModel2.userSession
-            print(viewModel2.userSession)
-        }
-       
             .toolbar {
-            
-            ToolbarItemGroup(placement: .bottomBar) {
                 
-                    Button("First") {
-                        self.value.toggle()
-                        viewModel.createRestaurant()
-                    }
+                ToolbarItemGroup(placement: .bottomBar) {
                     
-                
-                Button("Second") {
-                    withAnimation{
-                        viewModel2.logout()
-                        viewModel2.fetchUser()
+                        Button("First") {
+                           if viewModel2.userSession != nil  {
+                                viewModel.createRestaurant()
+                                viewModel2.currentUser?.value.toggle()
+                                viewModel2.EditProducti(id: viewModel2.currentUser?.id  ?? "reewwedsewd", value: viewModel2.currentUser?.value ?? true)
+                            }
+                        }
+                        .disabled( viewModel2.currentUser?.value == true)
+                    
+                    
+                    
+                    Button("Second") {
+                        withAnimation{
+                            viewModel2.logout()
+                            viewModel2.fetchUser()
+                        }
                     }
                 }
             }
-        }
-    
-        
+            
+            
+            
         
     }
 }
